\name{FrequencyListDataFrame}
\alias{FrequencyListDataFrame}
\alias{FrequencyListDataFrame-class}
\title{The FrequencyListDataFrame class}
\description{
  This class is the default implementation of
  \code{\link{FrequencyList}}. Considere the function
  \code{\link{readFrequencyList}} and
  \code{\link{writeFrequencyList}} for read/write access to
  files.
}
\author{
  Sylvain Loiseau
}
\seealso{
  \code{\link{FrequencyList}} ; \code{\link{Corpus}} for an
  overview of the available classes.
}

