\name{FullText}
\alias{FullText}
\alias{FullText-class}
\title{The parent of classes representing corpus as a list of character (tokens) vectors (= parts)}
\description{
  This virtual class is a subclass of
  \code{\link{CorpusAsTokens}} and the root hierarchy of
  all classes representing corpora as sequences of
  character strings.
}
\author{
  Sylvain Loiseau
}
\seealso{
  its superclass \code{\link{CorpusAsTokens}}; its sibling
  \code{\link{Tabulated}}; one of its implementation:
  \code{\link{FullTextList}}. \code{\link{Corpus}} for an
  overview of the available classes.
}

