\name{FullTextList}
\alias{FullTextList}
\alias{FullTextList-class}
\title{The FullTextList class}
\description{
  This class is the default implementation of
  \code{\link{FullText}}. Considere the function
  \code{\link{readTexts}} for creating easily
  \code{FullTextList} corpus thanks to text files.
}
\author{
  Sylvain Loiseau
}
\seealso{
  \code{\link{FullText}} ; \code{\link{Corpus}} for an
  overview of the available classes.
}

