\name{lexicalTable}
\alias{lexicalTable}
\alias{lexicalTable}
\title{Construct a \code{\link{LexicalTable}}}
\description{
  Construct a \code{\link{LexicalTable}}
}
\usage{
  lexicalTable(mat)
}
\arguments{
  \item{mat}{a matrix (data.frame, matrix or sparse matrix) containing the subfrequencies of
  types (in columns) in parts (in rows).}
}
\value{
  \code{\link{LexicalTable}}
} 
\seealso{
  \code{\link{LexicalTable}}
}
