\name{LexicalTableSparseMatrix}
\alias{LexicalTableSparseMatrix}
\alias{LexicalTableSparseMatrix-class}
\title{The LexicalTableSparseMatrix class}
\description{
  This class is the default implementation of
  \code{\link{LexicalTable}}.
}
\author{
  Sylvain Loiseau
}
\seealso{
  \code{\link{LexicalTable}} ; \code{\link{Corpus}} for an
  overview of the available classes.
}

