\name{readLexicalTable}
\alias{readLexicalTable}
\title{
read a \code{\link{LexicalTable}} previously saved with \code{\link{writeLexicalTable}}.
}
\description{
read a \code{\link{LexicalTable}} previously savec with \code{\link{writeLexicalTable}}.
}
\usage{
readLexicalTable(basename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{basename}{
     the name of a file created by \code{\link{writeLexicalTable}}, without its suffix.
}
}
\details{

}
\value{
A \code{\link{LexicalTable}} object.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{LexicalTable}}
}
\examples{
}
