\name{readTexts}
\alias{readTexts}

\title{
Read text files and create a \code{\link{FullText}} corpus.
}
\description{

}
\usage{
readTexts(filenames, is.dir = FALSE, pattern = NULL, split.on = "lines", enc = "UTF-8", skipEmpty = TRUE)
}

\arguments{
  \item{filenames}{

}
  \item{is.dir}{

}
  \item{pattern}{

}
  \item{split.on}{

}
  \item{enc}{

}
  \item{skipEmpty}{

}
}
\details{

}
\value{
A \code{\link{FullText}} corpus.
}
\references{

}
\author{

}
\note{

}

\seealso{

}
\examples{
}
