\name{specificites.lexicon}
\Rdversion{1.1}
\alias{specificites.lexicon}

\title{
specificites association score with two frequency lists.
}

\description{
Compute specificites association score between a lexicon and a sub-lexicon
}

\usage{
specificites.lexicon(lexicon, sublexicon)
}


\arguments{
  \item{lexicon}{ a frequency list (named vector) }
  \item{sublexicon}{ a frequency list (named vector) }
}

\value{
  specificites index as a named vector.
}

\seealso{
  \code{\link{specificites}} for specificites score and reference
}
