\name{tabulated}
\alias{tabulated}
\title{
Create a \code{\link{Tabulated}} Corpus.
}
\description{
}
\usage{
tabulated(m, structural.attributes = 0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{m}{
a data frame.
}
  \item{structural.attributes}{
The number of structural attribute (last column of the data frame. The n (n = structural.attribute)
last column of the data.frame contains numeric value (region id).
}
}
\details{
}
\value{
A \code{\link{Tabulated}} Corpus.
}
\references{
}
\author{
}
\note{
}
\seealso{
}
\examples{
}
