\name{TabulatedDataFrame}
\alias{TabulatedDataFrame}
\alias{TabulatedDataFrame-class}
\title{The TabulatedDataFrame class}
\description{
  This class is the default implementation of
  \code{\link{Tabulated}}. It can be used for representing
  corpus read from tabulated files as produced by most POS
  tagger.
}
\author{
  Sylvain Loiseau
}
\seealso{
  \code{\link{Tabulated}} ; \code{\link{Corpus}} for an
  overview of the available classes.
}

