\name{writeLexicalTable}
\alias{writeLexicalTable}
\title{
  write a \code{\link{LexicalTable}} into files.
}
\description{

}
\usage{
writeLexicalTable(obj, file)
}
\arguments{
  \item{obj}{
A \code{\link{LexicalTable}}
}
  \item{file}{
The file in which write the \code{\link{LexicalTable}}
}
}
\details{
See \code{\link{readLexicalTable}}
}
\value{
}
\references{

}
\author{

}
\note{

}

\seealso{
\code{\link{LexicalTable}}, \code{\link{readLexicalTable}}
}
\examples{
}
